breaking out inputs into separate stories
