+++ angular/projects/spark-angular/src/lib/components/inputs/Checkbox.stories.ts
@@ -1,0 +1,125 @@
+import { storyWrapper } from '../../../../../../.storybook/helpers/storyWrapper';
+import { SprkLabelModule } from '../../directives/inputs/sprk-label/sprk-label.module';
+import { SprkSelectionContainerModule } from './sprk-selection-container/sprk-selection-container.module';
+import { SprkSelectionItemContainerModule } from './sprk-selection-item-container/sprk-selection-item-container.module';
+import { SprkSelectionContainerComponent} from './sprk-selection-container/sprk-selection-container.component';
+import { SprkSelectionLabelModule } from '../../directives/inputs/sprk-selection-label/sprk-selection-label.module';
+
+export default {
+  title: 'Components/Input/Checkbox',
+  component: SprkSelectionContainerComponent,
+  decorators: [
+    storyWrapper(
+      storyContent => (
+        `<div class="sprk-o-Box">
+          <form (submit)="onSubmit($event)" #sampleForm="ngForm">
+            ${storyContent}
+          </form>
+        <div>`
+      )
+    )
+  ],
+  props: {
+    onSubmit(event): void {
+      this.form_submitted = true;
+    }
+  },
+  parameters: {
+    info: `
+##### For design and usage information check out the [documentation.](https://spark-docs.netlify.com/using-spark/components/input)
+    `,
+    docs: { iframeHeight: 200 },
+  }
+};
+
+const modules = {
+  imports: [
+    SprkLabelModule,
+    SprkSelectionContainerModule,
+    SprkSelectionItemContainerModule,
+    SprkSelectionLabelModule,
+  ],
+};
+
+export const checkbox = () => ({
+  moduleMetadata: modules,
+  template: `
+    <sprk-selection-container>
+      <label sprkLabel>
+        Checkbox Group Label
+      </label>
+
+      <sprk-selection-item-container>
+        <input
+          type="checkbox"
+          name="checkbox_input"
+          [(ngModel)]="checkbox_input1"
+          sprkSelectionInput
+          #checkboxInput1="ngModel"
+          data-id="checkbox-1"
+          id="checkbox-1"
+          (change)="onSelect($event)"
+        >
+
+        <label
+          for="checkbox-1"
+          sprkSelectionLabel
+        >
+          Checkbox Item 1
+        </label>
+      </sprk-selection-item-container>
+
+      <sprk-selection-item-container>
+        <input
+          type="checkbox"
+          name="checkbox_input"
+          [(ngModel)]="checkbox_input2"
+          sprkSelectionInput
+          #checkboxInput2="ngModel"
+          data-id="checkbox-2"
+          id="checkbox-2"
+          (change)="onSelect($event)"
+        >
+
+        <label
+          for="checkbox-2"
+          sprkSelectionLabel
+        >
+          Checkbox Item 2
+        </label>
+      </sprk-selection-item-container>
+
+      <sprk-selection-item-container>
+        <input
+          type="checkbox"
+          name="checkbox_input"
+          [(ngModel)]="checkbox_input3"
+          sprkSelectionInput
+          #checkboxInput3="ngModel"
+          data-id="checkbox-3"
+          id="checkbox-3"
+          (change)="onSelect($event)"
+        >
+
+        <label
+          for="checkbox-3"
+          sprkSelectionLabel
+        >
+          Checkbox Item 3
+        </label>
+      </sprk-selection-item-container>
+    </sprk-selection-container>
+  `,
+  props: {
+    onSelect() {
+      this.checkbox_input1 === true
+        ? (this.isChecked = true)
+        : (this.isChecked = false);
+    }
+  },
+});
+
+checkbox.story = {
+  name: 'Default',
+};
+
