+++ angular/projects/spark-angular/src/lib/components/inputs/Date.stories.ts
@@ -1,0 +1,74 @@
+import { storyWrapper } from '../../../../../../.storybook/helpers/storyWrapper';
+import { SparkInputContainerModule } from './sprk-input-container/sprk-input-container.module';
+import { SprkInputModule } from '../../directives/inputs/sprk-input/sprk-input.module';
+import { SprkIconModule } from '../sprk-icon/sprk-icon.module';
+import { SprkLabelModule } from '../../directives/inputs/sprk-label/sprk-label.module';
+import { SprkFieldErrorModule } from '../../directives/inputs/sprk-field-error/sprk-field-error.module';
+import { SprkIconInputContainerComponent } from './sprk-icon-input-container/sprk-icon-input-container.component';
+
+export default {
+  title: 'Components/Input/Date',
+  component: SprkIconInputContainerComponent,
+  decorators: [
+    storyWrapper(
+      storyContent => (
+        `<div class="sprk-o-Box">
+          <form (submit)="onSubmit($event)" #sampleForm="ngForm">
+            ${storyContent}
+          </form>
+        <div>`
+      )
+    )
+  ],
+  props: {
+    onSubmit(event): void {
+      this.form_submitted = true;
+    }
+  },
+  parameters: {
+    info: `
+##### For design and usage information check out the [documentation.](https://spark-docs.netlify.com/using-spark/components/input)
+    `,
+    docs: { iframeHeight: 200 },
+  }
+};
+
+const modules = {
+  imports: [
+    SparkInputContainerModule,
+    SprkLabelModule,
+    SprkInputModule,
+    SprkIconModule,
+    SprkFieldErrorModule,
+  ],
+};
+
+export const dateInput = () => ({
+  moduleMetadata: modules,
+  template: `
+    <sprk-input-container>
+      <label sprkLabel>Date Input (No Picker)</label>
+      <input
+        name="date_input"
+        type="text"
+        pattern="^(((0[13578]|1[02])([\\/-]?)(0[1-9]|[12]\\d|3[01])|(0[469]|11)([\\/-]?)(0[1-9]|[12]\\d|30)|02([\\/-]?)((0[1-9])|[12]\\d))(\\4|\\7|\\9)[12]\\d{3})?$"
+        placeholder="MM/DD/YYYY"
+        [(ngModel)]="date_input"
+        #dateInput="ngModel"
+        sprkFormatterDate
+        sprkInput
+      />
+      <span [hidden]="dateInput.valid || dateInput.pristine" sprkFieldError>
+        <sprk-icon
+          iconType="exclamation-filled-small"
+          additionalClasses="sprk-b-ErrorIcon"
+        ></sprk-icon>
+        <div class="sprk-b-ErrorText">Invalid date.</div>
+      </span>
+    </sprk-input-container>
+  `,
+});
+
+dateInput.story = {
+  name: 'Default',
+};
